+++ main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"fairy-tale-generator/amazonpolly"
 	"fairy-tale-generator/openai"
+	"fairy-tale-generator/story"
 	"flag"
 	"fmt"
 	"log"
-	"math/rand"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -23,7 +23,7 @@ import (
 )
 
 var (
-	// Version is the version of the program.
+	// Version of this program, set during compile time.
 	Version = "dev"
 )
 
@@ -36,6 +36,7 @@ const (
 	model = "gpt-3.5-turbo"
 )
 
+// Secrets for the APIs
 var (
 	apiKey             = os.Getenv("OPENAI_API_KEY")
 	orgID              = os.Getenv("OPENAI_ORGANIZATION")
@@ -43,32 +44,8 @@ var (
 	awsSecretAccessKey = os.Getenv("AWS_SECRET_ACCESS_KEY")
 )
 
-// Embeded text files
 var (
-	//go:embed asssets/charactersupporterset.txt
-	characterSupporterSetAsset string
-	//go:embed asssets/locationset.txt
-	locationSetAsset string
-	//go:embed asssets/storyplotsset.txt
-	storyPlotsSetAsset string
-	//go:embed asssets/charactermainset.txt
-	characterMainSetAsset string
-)
-
-// Options for the fairy tale
-var (
-	// CharacterSupporterSet is the set of supporter characters for the fairy tale.
-	CharacterSupporterSet = textToLinesInRandomOrder(characterSupporterSetAsset)
-	// LocationSet is the set of locations for the fairy tale.
-	LocationSet = textToLinesInRandomOrder(locationSetAsset)
-	// StoryPlotSet is the set of story plots for the fairy tale.
-	StoryPlotSet = textToLinesInRandomOrder(storyPlotsSetAsset)
-	// CharacterMainSet is the set of main characters for the fairy tale.
-	CharacterMainSet = textToLinesInRandomOrder(characterMainSetAsset)
-)
-
-var (
-	flatSetRandomOptions = flag.Bool("random", false, "Use a flat set of random options instead of the default interactive options")
+	flatSetRandomOptions = flag.Bool("random", false, "Use a set of random options for the strory instead of user interactive input")
 )
 
 var Commit = func() string {
@@ -98,7 +75,7 @@ func main() {
 	fmt.Println()
 	fmt.Println("This program uses OpenAI's GPT-3 API to generate a fairy tale in German with the help of AWS Polly.")
 	fmt.Println()
-	pterm.Info.Printf("Loaded %v main charaters, %v support charaters, %v locations, %v plots which results in %v possible stories\n", len(CharacterMainSet), len(CharacterSupporterSet), len(LocationSet), len(StoryPlotSet), len(CharacterMainSet)*len(CharacterSupporterSet)*len(LocationSet)*len(StoryPlotSet))
+	pterm.Info.Printf("Loaded %v main charaters, %v support charaters, %v locations, %v plots which results in %v possible stories\n", len(story.CharacterMainSet), len(story.CharacterSupporterSet), len(story.LocationSet), len(story.StoryPlotSet), len(story.CharacterMainSet)*len(story.CharacterSupporterSet)*len(story.LocationSet)*len(story.StoryPlotSet))
 	fmt.Println()
 
 	flag.Parse()
@@ -115,7 +92,7 @@ func main() {
 		return
 	}
 
-	fairyTaleOptions := getFairyTaleOptions(*flatSetRandomOptions, ChapterCount)
+	fairyTaleOptions := story.GetFairyTaleOptions(*flatSetRandomOptions, ChapterCount)
 
 	openai := openai.New(apiKey, orgID, model, fairyTaleOptions)
 	amazonpolly := amazonpolly.New(awsKeyId, awsSecretAccessKey)
@@ -160,21 +137,6 @@ func generateAndPlay(targetFolder string, openai *openai.OpenAI, amazonpolly *am
 			return
 		}
 	}
-}
-
-func textToLinesInRandomOrder(text string) []string {
-	lines := []string{}
-	for _, line := range strings.Split(text, "\n") {
-		if line != "" {
-			lines = append(lines, strings.Trim(line, ""))
-		}
-	}
-
-	rand.Shuffle(len(lines), func(i, j int) {
-		lines[i], lines[j] = lines[j], lines[i]
-	})
-
-	return lines
 }
 
 func createTimestamp() string {
