+++ story/asssets/charactermainset.txt
