+++ story/asssets/charactersupporterset.txt
