+++ story/asssets/locationset.txt
